Several updates and bug fixes
Added an idle right animation so the player isn't always facing left. Added sound effects to the desert world. Made it so you can select the world on the map using the keyboard. Cleaned up some code. Added a quest indicator to the quest npc. Started implementing a quest list on the side of the screen before i realized I couldn't understand Ethan's code... Added an indicator on the world map to see what room you are in.