Fixed some bugs. Made some graphical changes.
Fixed the really loud audio bug by adjusting the available volume levels.
Added a percentage counter for the volume slider.
Added the volume title for the settings menu.
Created a save indicator for the save NPC.
Added the volume menu to the main menu so the player can adjust the volume before they get stuck in the tutorial.
Made it so you can see what treasure you need to get in the quest menu.
Created and added a new less blurry game saved asset.
The questlist was showing up in the main menu when the player went to the main menu from the pause menu, so i made it hide itself in the main menu scene.